Fixing PR comments
