+++ Content.Client/Devour/DevourSystem.cs
@@ -1,0 +1,6 @@
+using Content.Shared.Devour;
+
+namespace Content.Client.Devour;
+public sealed class DevourSystem : SharedDevourSystem
+{
+}
