+++ Content.Shared/Devour/Components/DevourerComponent.cs
@@ -5,11 +5,12 @@ using Content.Shared.Devour;
 using Content.Shared.Whitelist;
 using Robust.Shared.Audio;
 using Robust.Shared.Containers;
+using Robust.Shared.GameStates;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype;
 
 namespace Content.Server.Devour.Components;
 
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent]
 [Access(typeof(SharedDevourSystem))]
 public sealed class DevourerComponent : Component
 {
