--- Content.Shared/Devour/DevourActionEvent.cs
@@ -1,5 +1,0 @@
-using Content.Shared.Actions;
-
-namespace Content.Shared.Devour;
-
-public sealed class DevourActionEvent : EntityTargetActionEvent { }
