--- Content.Shared/Devour/DevourDoAfterEvent.cs
@@ -1,7 +1,0 @@
-using Content.Shared.DoAfter;
-using Robust.Shared.Serialization;
-
-namespace Content.Shared.Devour;
-
-[Serializable, NetSerializable]
-public sealed class DevourDoAfterEvent : SimpleDoAfterEvent { }
